requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --output-file=requirements.txt --pip-args=None --strip-extras requirements.in
#
--index-url https://download.pytorch.org/whl/cpu
--extra-index-url https://pypi.org/simple

albumentations==1.3.1
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
antlr4-python3-runtime==4.9.3
    # via omegaconf
filelock==3.20.0
    # via torch
fsspec==2025.10.0
    # via torch
imageio==2.37.2
    # via scikit-image
jinja2==3.1.6
    # via torch
joblib==1.5.2
    # via scikit-learn
lazy-loader==0.4
    # via scikit-image
markupsafe==3.0.3
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via
    #   scikit-image
    #   torch
numpy==2.2.6
    # via
    #   albumentations
    #   imageio
    #   opencv-python
    #   opencv-python-headless
    #   pandas
    #   qudida
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   tifffile
    #   torchvision
omegaconf==2.3.0
    # via -r requirements.in
opencv-python==4.12.0.88
    # via -r requirements.in
opencv-python-headless==4.12.0.88
    # via
    #   albumentations
    #   qudida
packaging==25.0
    # via
    #   lazy-loader
    #   scikit-image
pandas==2.3.3
    # via -r requirements.in
pillow==12.0.0
    # via
    #   imageio
    #   scikit-image
    #   torchvision
pydantic==2.12.3
    # via -r requirements.in
pydantic-core==2.41.4
    # via pydantic
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   albumentations
    #   omegaconf
qudida==0.0.4
    # via albumentations
scikit-image==0.25.2
    # via albumentations
scikit-learn==1.7.2
    # via qudida
scipy==1.16.3
    # via
    #   albumentations
    #   scikit-image
    #   scikit-learn
simsimd==6.5.3
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
stringzilla==4.2.3
    # via -r requirements.in
sympy==1.14.0
    # via torch
threadpoolctl==3.6.0
    # via scikit-learn
tifffile==2025.10.16
    # via scikit-image
torch==2.9.0+cpu
    # via
    #   -r requirements.in
    #   torchaudio
    #   torchvision
torchaudio==2.9.0+cpu
    # via -r requirements.in
torchvision==0.24.0+cpu
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   pydantic
    #   pydantic-core
    #   qudida
    #   torch
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas

# The following packages are considered to be unsafe in a requirements file:
# setuptools
